arabic_stt/requirements.txt
# Deep Learning
torch>=2.0.0
torchaudio>=2.0.0

# Audio Processing
librosa>=0.10.0
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.24.0

# Arabic Text Processing
arabic-reshaper>=3.0.0
python-bidi>=0.4.2

# Data Handling
pandas>=2.0.0
datasets>=2.14.0  # For loading Common Voice dataset

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
tensorboard>=2.13.0

# Real-time inference
pyaudio>=0.2.13
webrtcvad>=2.0.10  # Voice Activity Detection

# ONNX Export (for lightweight deployment)
onnx>=1.14.0
onnxruntime>=1.15.0
onnx-simplifier>=0.4.0  # Optional: simplify exported models
